codecs: nvv4l2: support multiple L4T versions with single build
Nvidia has the tendency to break compatibility on documented apis for no reason.

Adding new values to enums in-between older values and not the end is unacceptable.

A heuristic is now in use that can identify NvBuffer version and adjusts such values accordingly.

That depends on using nvv4l2_ext_utils header with the specific values there.