codecs: nvv4l2: align line width to 64B
Transformations of formats to formats of simply Block linear to Pitch are done in HW.

VIC is responsible for that and has the limitation that width must be aligned to 64 Bytes.

This allows the decoder to work with every single non standard resolution, effectively bringing full support to any video.

That is particularly important on emulators that use ffmpeg for decoding in game videos which sometimes are used as backgrounds.
These have weird dimensions and need that fix.

The output frame of course has the original line width.