codecs: nvv4l2: fix hanging on event wait if no full frame info
Sometimes the decoder can be opened with no packet sent or not many for a frame to be decoded.

In such cases the thread will get stuck for 50s waiting for a resolution event to happen.

Use the streamon and eos checks properly to get out of such a loop.