codecs: nvv4l2: align encoder plane width to 64B
Transformations of Block linear formats to linear Pitch formats are done in HW.

VIC is responsible for that and has the limitation that width must be aligned to 64 Bytes.

This allows the encoder to work with every single non standard resolution, effectively bringing full support to any video.

That is particularly important on streaming software that use ffmpeg for encoding which many times can be weird resolutions (e.g. pip).

The output encoded frame has the original line width of course.