codecs: nvv4l2: align enc plane width per format/plane
Take two on creating a heuristic of the needed alignment.

Unlike in VIC/NVDEC, in encoding the conversion is from Pitch to Block linear.
That needs precise alignments for each plane, per format.

This supports way more resolutions than before, but still lacks support for non-standard ones.

For example, 854 width needs 32B alignment (instead of 64!) for yuv420 on main plane.
And the other 2 planes need 64B.

As usual, TRM is not helpful on that (only has 64/256B alignments for block/pitch)
and there's probably an algorithm on the drivers that causes that issue by aligning
sizes for NVENC without notifying user.