nvv4l2: allow 10-bit HEVC
The only hw support for 10-bit on NVDEC 2nd/3rd gen is on HEVC. H264 does not support it on main spec, so as per previous changes it fallbacks to software decoding. VP8 codec does not support it at all. The only out of spec non-support on hw decoding is on VP9.